Optimise the messages

+++ DatingApp/API/SignalR/MessageHub.cs
@@ -16,14 +16,16 @@ public class MessageHub(IMessageRepository messageRepository, IUserRepository us
         if (Context.User == null || string.IsNullOrEmpty(otherUser)) throw new Exception("Cannot join group");
         var groupName = GetGroupName(Context.User.GetUsername(), otherUser);
         await Groups.AddToGroupAsync(Context.ConnectionId, groupName);
-        await AddToGroup(groupName);
+        var group = await AddToGroup(groupName);
+        await Clients.Group(groupName).SendAsync("UpdatedGroup", group);
         var messages = await messageRepository.GetMessageThread(Context.User.GetUsername(), otherUser!);
-        await Clients.Group(groupName).SendAsync("ReceivedMessageThread", messages);
+        await Clients.Caller.SendAsync("ReceivedMessageThread", messages);
     }
 
     public override async Task OnDisconnectedAsync(Exception? exception)
     {
-        await RemoveFromMessageGroup();
+        var group = await RemoveFromMessageGroup();
+        await Clients.Group(group.Name).SendAsync("UpdatedGroup", group);
         await base.OnDisconnectedAsync(exception);
     }
 
@@ -76,7 +78,7 @@ public class MessageHub(IMessageRepository messageRepository, IUserRepository us
         }
     }
 
-    public async Task<bool> AddToGroup(string groupName)
+    public async Task<Group> AddToGroup(string groupName)
     {
         var username = Context.User?.GetUsername() ?? throw new Exception("Cannot get username");
         var group = await messageRepository.GetMessageGroup(groupName);
@@ -90,17 +92,23 @@ public class MessageHub(IMessageRepository messageRepository, IUserRepository us
 
         group.Connections.Add(connection);
 
-        return await messageRepository.SaveAllAsync();
+        if (await messageRepository.SaveAllAsync()) return group;
+
+        throw new HubException("Failed to joing group");
     }
 
-    private async Task RemoveFromMessageGroup()
+    private async Task<Group> RemoveFromMessageGroup()
     {
-        var connection = await messageRepository.GetConnection(Context.ConnectionId);
-        if (connection != null)
+        var group = await messageRepository.GetGroupForConnection(Context.ConnectionId);
+        var connection = group?.Connections.FirstOrDefault(x => x.ConnectionId == Context.ConnectionId);
+
+        if (connection != null && group != null)
         {
             messageRepository.RemoveConnection(connection);
-            await messageRepository.SaveAllAsync();
+            if (await messageRepository.SaveAllAsync()) return group;
         }
+
+        throw new HubException("Failed to remove from group");
     }
 
     private string GetGroupName(string caller, string? other)
